bug fixed

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,23 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
-        Image image = imageService.getImage(imageId);
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+    	User user = (User) session.getAttribute("loggeduser");
+    	Image image = imageService.getImage(imageId);
 
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if(user.getId()==image.getUser().getId())
+        {
+        	 model.addAttribute("tags", tags);
+             return "images/edit";
+        }
+        else
+        {
+        	model.addAttribute("tags", image.getTags());
+        	model.addAttribute("editError", true);
+        	return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +150,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+    	Image image = imageService.getImage(imageId);
+    	User user = (User) session.getAttribute("loggeduser");
+    	if(user.getId()==image.getUser().getId())
+    	{	
+    	imageService.deleteImage(imageId);
         return "redirect:/images";
+    	}
+    	else
+    	{
+    		model.addAttribute("image", image);
+    		model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError", true);
+            return "images/image";
+    	}
     }
